+++ functions/functions.php
@@ -8,12 +8,14 @@ use Laminas\Config\Factory;
 use Laminas\Stdlib\Glob;
 
 use function getenv;
+use function implode;
+use function is_array;
+use function is_scalar;
+use function putenv;
+use function sprintf;
 use function strtolower;
 use function trim;
 
-/**
- * Loads configuration files which match provided glob pattern, and returns the merged result as array
- */
 function loadConfigFromGlob(string $globPattern): array
 {
     /** @var array $config */
@@ -35,4 +37,19 @@ function env(string $key, mixed $default = null): mixed
         'null', '(null)' => null,
         default => trim($value),
     };
+}
+
+function putNotYetDefinedEnv(string $key, mixed $value): void
+{
+    $isArray = is_array($value);
+    if (!($isArray || is_scalar($value)) || env($key) !== null) {
+        return;
+    }
+
+    $normalizedValue = $isArray ? implode(',', $value) : match ($value) {
+        true => 'true',
+        false => 'false',
+        default => (string) $value,
+    };
+    putenv(sprintf('%s=%s', $key, $normalizedValue));
 }